1.change the toast of copying Multi-Media message or WAP Push message to card;2.do not delete locked messages in detail screen

+++ src/com/android/mms/ui/ComposeMessageActivity.java
@@ -902,7 +902,8 @@ public class ComposeMessageActivity extends Activity
         if (MessageUtils.getActivatedIccCardCount() > 1) {
             int preferredSub = getPreferredSubscription();
             boolean alwaysAsk = (preferredSub != MSimConstants.SUB1 && preferredSub != MSimConstants.SUB2);
-
+            Log.v(TAG,"preferredSub = " + preferredSub + ", alwaysAsk = " + alwaysAsk
+                 + ", mSendSubscription = " + mSendSubscription);
             if (alwaysAsk && mSendSubscription == SUBSCRIPTION_ID_INVALID) {
                 if (mChooseDialog == null || !mChooseDialog.isShowing()) {
                     LaunchChooseDialog(bCheckEcmMode, isMms);
@@ -918,8 +919,8 @@ public class ComposeMessageActivity extends Activity
                 sendMessage(bCheckEcmMode);
             }
         } else {
-            int preferredSmsSub = MSimSmsManager.getDefault().getPreferredSmsSubscription();
-            if(preferredSmsSub == ALWAY_ASK){
+            //int preferredSmsSub = MSimSmsManager.getDefault().getPreferredSmsSubscription();
+            /*if(preferredSmsSub == ALWAY_ASK){*/
                 int availableSub = getAvailableSub();
                 if(availableSub != -1){
                     mWorkingMessage.setCurrentConvSub(availableSub);
@@ -928,11 +929,11 @@ public class ComposeMessageActivity extends Activity
                     Toast.makeText(ComposeMessageActivity.this,R.string.cannot_send_message, Toast.LENGTH_LONG).show();
                     return;
                 }
-            }else{
+            /*}else{
                 mWorkingMessage.setCurrentConvSub(MSimSmsManager.getDefault()
                         .getPreferredSmsSubscription());
                 sendMessage(bCheckEcmMode);
-            }
+            }*/
         }
     }
 
@@ -1748,6 +1749,12 @@ public class ComposeMessageActivity extends Activity
                     return showMessageDetails(mMsgItem);
 
                 case MENU_DELETE_MESSAGE: {
+                    if(msgItem.mLocked)
+                    {
+                        Toast.makeText(ComposeMessageActivity.this, R.string.delete_lock_err, Toast.LENGTH_LONG).show();
+                        return true;
+                    }
+                                        
                     DeleteMessageListener l = new DeleteMessageListener(mMsgItem);
                     confirmDeleteDialog(l, mMsgItem.mLocked);
                     return true;
